+++ examples/scenes.rs
@@ -55,13 +55,13 @@ enum SceneResult {
 }
 
 impl Scene<SceneResult, SceneName> for Scene1 {
-    fn render(&self, graphics: &mut Graphics, _: Coord) {
+    fn render(&self, graphics: &mut Graphics, _: Coord, _: &[KeyCode], _: &GameController) {
         graphics.clear(BLUE);
         graphics.set_pixel(0, 0, RED);
         graphics.set_pixel(1, 1, RED);
     }
 
-    fn on_mouse_up(&mut self, _: Coord, button: MouseButton, _: &Vec<&KeyCode>) {
+    fn on_mouse_up(&mut self, _: Coord, button: MouseButton, _: &[KeyCode]) {
         if button != MouseButton::Left {
             return;
         }
@@ -72,7 +72,8 @@ impl Scene<SceneResult, SceneName> for Scene1 {
         &mut self,
         _: &Timing,
         _: Coord,
-        _: &Vec<&KeyCode>,
+        _: &[KeyCode],
+        _: &GameController,
     ) -> SceneUpdateResult<SceneResult, SceneName> {
         self.result.clone()
     }
@@ -89,15 +90,15 @@ impl Scene<SceneResult, SceneName> for Scene1 {
 }
 
 impl Scene<SceneResult, SceneName> for Scene2 {
-    fn render(&self, graphics: &mut Graphics, _: Coord) {
+    fn render(&self, graphics: &mut Graphics, _: Coord, _: &[KeyCode], _: &GameController) {
         graphics.clear(RED);
     }
 
-    fn on_key_up(&mut self, key: KeyCode, _: Coord, _: &Vec<&KeyCode>) {
+    fn on_key_up(&mut self, key: KeyCode, _: Coord, _: &[KeyCode]) {
         self.result = Pop(Some(FromKey(key)))
     }
 
-    fn on_mouse_up(&mut self, xy: Coord, button: MouseButton, _: &Vec<&KeyCode>) {
+    fn on_mouse_up(&mut self, xy: Coord, button: MouseButton, _: &[KeyCode]) {
         if button != MouseButton::Left {
             return;
         }
@@ -112,7 +113,8 @@ impl Scene<SceneResult, SceneName> for Scene2 {
         &mut self,
         _: &Timing,
         _: Coord,
-        _: &Vec<&KeyCode>,
+        _: &[KeyCode],
+        _: &GameController,
     ) -> SceneUpdateResult<SceneResult, SceneName> {
         self.result.clone()
     }
@@ -123,11 +125,11 @@ impl Scene<SceneResult, SceneName> for Scene2 {
 }
 
 impl Scene<SceneResult, SceneName> for Scene3 {
-    fn render(&self, graphics: &mut Graphics, _: Coord) {
+    fn render(&self, graphics: &mut Graphics, _: Coord, _: &[KeyCode], _: &GameController) {
         graphics.draw_rect(Rect::new((90, 90), (190, 150)), fill(self.back));
     }
 
-    fn on_mouse_up(&mut self, _: Coord, button: MouseButton, _: &Vec<&KeyCode>) {
+    fn on_mouse_up(&mut self, _: Coord, button: MouseButton, _: &[KeyCode]) {
         if button != MouseButton::Left {
             return;
         }
@@ -138,7 +140,8 @@ impl Scene<SceneResult, SceneName> for Scene3 {
         &mut self,
         _: &Timing,
         _: Coord,
-        _: &Vec<&KeyCode>,
+        _: &[KeyCode],
+        _: &GameController,
     ) -> SceneUpdateResult<SceneResult, SceneName> {
         self.result.clone()
     }